fix(preflight): exempt informational checks from the overall rollup
rollup() was counting a check's non-READY/NOT_APPLICABLE State as a
problem regardless of its Severity, contradicting SeverityInfo's own
documented contract ("never affects the overall rollup by itself") -
caught by TestBuildReport_FullyHealthy and TestBuildReport_
OptionalSensorDisabled, both of which include the fixed fan_rotation
VERIFY/INFO notice and unexpectedly saw it pull an otherwise-fully-ready
report to CAUTION.

Also gofmt's the whole package.

+++ preflight/checks.go
@@ -154,7 +154,7 @@ func gpsTimeChecks(in Input) []CheckResult {
 	// never itself a separate blocking/caution trigger.
 	if hasFix {
 		out = append(out, newCheck("GPS", "gps_accuracy", "GPS position accuracy", StateReady, SeverityInfo,
-			formatFloat(gps.AccuracyMeters)+" m estimated accuracy"))
+			formatFloat(float64(gps.AccuracyMeters))+" m estimated accuracy"))
 	} else {
 		out = append(out, newCheck("GPS", "gps_accuracy", "GPS position accuracy", StateUnknown, SeverityInfo, "no current position solution"))
 	}
